APP DEEPSEEK/requirements.txt
deepface==0.0.79
opencv-python==4.9.0.80
Pillow==10.1.0
numpy==1.24.3
tensorflow==2.15.0
tf-keras==2.15.0